Correction for #150 and #161. This change allows for icosa to be accessed as a subfolder of the domain, such as https://example.com/icosa.
The main modifications were to move the urls definition from the django_project to the icosa app, and to add a BASE_URL setting in settings.py.
A namespace was added to the urls in the icosa app, so that they can be accessed as icosa:asset_view, for example.
All the references to the urls in the templates and forms were updated to use this new namespace and the `url` template tag.
The api urls were also updated to use the new namespace and can be accessed as icosa:api:asset_list, for example.

This is quite a large change, but it is mostly just moving code around and updating references. However, it does require some testing to ensure that everything works as expected.